+++ deploy/complete/terraform/outputs.tf
@@ -21,7 +21,7 @@ output "mushop_url_https" {
   depends_on = [helm_release.ingress_nginx]
 }
 output "grafana_url" {
-  value       = format("${local.mushop_url_protocol}://%s/grafana", local.mushop_ingress_hostname)
+  value       = var.grafana_enabled ? format("${local.mushop_url_protocol}://%s/grafana", local.mushop_ingress_hostname) : null 
   description = "Grafana Dashboards URL"
 
   depends_on = [helm_release.ingress_nginx]
@@ -41,7 +41,7 @@ output "autonomous_database_password" {
   value = random_string.autonomous_database_admin_password.result
 }
 output "grafana_admin_password" {
-  value = nonsensitive(data.kubernetes_secret.mushop_utils_grafana.data.admin-password) # Required for TF 0.15, as automatically generate an error if is not marked as sensitive
+  value = var.grafana_enabled ? nonsensitive(data.kubernetes_secret.mushop_utils_grafana.0.data.admin-password) : null # Required for TF 0.15, as automatically generate an error if is not marked as sensitive
   # value     = data.kubernetes_secret.mushop_utils_grafana.data.admin-password # TF 0.14 version support by ORM does not support nonsensitive function
 }
 output "mushop_source_code" {
@@ -51,4 +51,8 @@ locals {
   mushop_ingress_ip       = data.kubernetes_service.mushop_ingress.load_balancer_ingress[0].ip
   mushop_ingress_hostname = var.ingress_hosts == "" ? data.kubernetes_service.mushop_ingress.load_balancer_ingress[0].ip : split(",", var.ingress_hosts)[0]
   mushop_url_protocol     = var.ingress_tls ? "https" : "http"
+}
+
+output "temp" {
+  value = oci_apigateway_deployment.newsletter_subscription.endpoint
 }